+++ genairics/__main__.py
@@ -10,7 +10,7 @@ def prepareParser():
     """
     import argparse
     from collections import OrderedDict
-    from genairics import typeMapping
+    from genairics import typeMapping, config
     from genairics.jobs import QueuJob
     from genairics.RNAseq import RNAseq
     from genairics.ChIPseq import ChIPseq
@@ -68,6 +68,9 @@ def prepareParser():
                         help='choose where and how the job will run')
     parser.add_argument('--cluster-Q', default = '', type = str,
                         help = 'the queue and/or job cluster to which job will be submitted. Format should be qname@clusterservername')
+    parser.add_argument('--cluster-PPN', default = config.threads, type = int,
+                        help = '''Number of processors per node that will be requested 
+(recommended > 12) (default: {})'''.format(config.threads))
     parser.add_argument('--remote-host', default = '', type = str, help = 'submit job through ssh')
     parser.add_argument('--save-config', action = 'store_true',
                         help = 'save path related default values to a configuration file in the directory where you started genairics')
@@ -168,6 +171,7 @@ def main(args=None):
     joblauncher = joblaunchers[args.pop('job_launcher')]
     remotehost = args.pop('remote_host')
     clusterQ = args.pop('cluster_Q')
+    clusterPPN = args.pop('cluster_PPN')
     verbose = args.pop('verbose')
     workflow = args.pop('function')(**args)
 
@@ -177,7 +181,10 @@ def main(args=None):
     
     if joblauncher:
         logger.debug('submitting %s to %s',workflow,joblauncher)
-        joblauncher(job=workflow,remote=remotehost,clusterQ=clusterQ).run()
+        joblauncher(
+            job=workflow,remote=remotehost,
+            clusterQ=clusterQ,clusterPPN=clusterPPN
+        ).run()
     else: runWorkflow(workflow)
 
 # Run main program logics when script called directly