+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project are documented in this file.
 
 ## [Unreleased]
+### Added
+- --cluster-PPN option for qsub job-launcher to specify number of
+processors per node to request.
 
 ## [0.2.1] - 2018-06-19
 ### Added